+++ src/main/java/ru/jsft/gtdfan/repository/TaskRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.jsft.gtdfan.model.Task;
 
 @Repository
-public interface TaskRepository extends CrudRepository<Task, Integer> {
+public interface TaskRepository extends CrudRepository<Task, Long> {
 }
